requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==42.0.1
daphne==4.0.0
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
MouseInfo==0.1.3
msgpack==1.0.7
phonenumbers==8.13.29
pillow==10.2.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
pyOpenSSL==24.0.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.0.7
pytz==2023.3.post1
redis==5.0.1
rest-framework-simplejwt==0.0.2
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
zope.interface==6.1
